visual_safety_requirements.txt
# Visual Safety Assessment Requirements
# Core ML/Vision dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0  # Still needed for Qwen-VL

# HuggingFace ecosystem
transformers>=4.37.0
accelerate>=0.25.0
datasets>=2.14.0
tokenizers>=0.15.0

# Vision-specific packages
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Evaluation and metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
wandb>=0.16.0
jinja2>=3.1.0
pyyaml>=6.0

# Optional: for advanced evaluation
# evaluate>=0.4.0  # HuggingFace evaluation library
# bert-score>=0.3.13  # For text evaluation if needed
